Rewrite the sequencefile to be faster and more memory-conscious
This churns all the sequencefile code and some of the index code, mainly for
the reasons of optimization. In particular, allocating a SectionReader for
each key/value was actually pretty disastrous for memory. The new approach
looks more like a bufio.Scanner, and, internally, reuses a single memory slice.

There's some other miscellaneous cleanup.