Fix race condition when switching versions
Previously, we didn't do any synchronization on the index, so when we
upgrade to a new index, the old one just gets stomped, even if there are
requests that are currently using that index.

The strategy for fixing this is as follows: we take out an RW mutex when
getting a reference to the index, and bump a refcount for that mutex.
When updating to a new index version, the upgrader takes out a write
lock on the mutex, swaps them, release the mutex, and then waits for the
refcount of the mutex to drain to 0 before destroying it. This is
reasonably efficient, since we only need mutual exclusion when getting
a reference to the index, not when actually using it. The logic for this
is mostly factored away in side of the newly created IndexReference
struct.