Add support for gzip encoding
We now look at the `Accept-Encoding` header to allow the client to
request gzip encoding.
Currently, the solution only considers the request to determine the
content encoding. In the future, we may want to augment this so sequins
can reply with the identity encoding for small requests, where gzipping
the request is likely to be counter-productive.