Introduce per-file indexing instead of a big global index
By default, if you produce a dataset using a Hadoop job with a reduce
step, your files will be sorted within the individual output files,
and partitioned based on the hashCode of the keys (as long as the
reduce key is the same as the output key).

This enables two interesting optimizations, the first of which is
implemented in this commit. When data is sorted, the index can be
sparse instead of complete; it can include a subset of the keys, for
spaced-out offsets in the file. Since this index is much smaller,
it can be easily stored in memory.

Then on a get, instead of seeking to exactly where the key is, we can
seek to the closest key before the target key, and scan the file
until we either find the target key or pass it.

While this degrades read performance very slightly in some cases,
it can also improve it in others, because the index lookup happens
completely in memory. And the indexing performance is improved
*substantially*, in many cases by several orders of magnitude.

The new indexing code tries to construct a sparse index for each file,
but will fall back to a total index if it detects the data isn't
sorted. Both types of indexes implement the `fileIndex` interface,
which is convenient for testing, etc.

The biggest downside to this approach is that, in the worst case,
we have to query the index of each file, one at a time, to find our
key. This can be mitigated by being aware of the partitioning of the
data, either detecting the hashCode partition of the data (which is
not implemented yet) or using the minimum and maximum key, if the data
is descretely partitioned into ranges of keys (which is implemented).

A few small changes snuck in as well; namely new, bigger test data and
the removal of the CRC check when loading from a manifest (which was
much too slow, and has been replaced with a simple file size check).