Allow nodes to join an existing cluster immediately
In cases where a node is joining a cluster that already has a version
ready to go, it should be able to start up and immediately start
serving requests to that version (but exclusively proxying them).

This has the happy side-effect of making the version interface a
little simpler, as well.