Swap in go-zookeeper for gozk
go-zookeeper seems to be giving us trouble. gozk is a bit old, and
perhaps unmaintained, but it's also only ~1000 lines of go; it wraps
the C client. To use it, we have to do this whole vendoring thing
again.

This also makes some cosmetic changes to the logging in zk-watcher,
and inverts the run() loop (since it's a bit cleaner this way).
Finally, it sneaks in some logging of peers, to aid in debugging.