Do our best to return data for misproxied requests
Under the current behavior, sequins will 501 if asked for a version
of a dataset it doesn't have. This is great from a correctness
standpoint - it means that we can gaurantee a single sequins node
will produce a consistent view, with no version slippage - but it
means that when sharding breaks for any other reason, the results can
be pretty dramatic.

It's important to emphasize availability in this (rare) case. Most
clients don't care if they see version A -> B -> A - they just want
some data. And we return the version with the data, so clients
can independently verify they're getting the right version anyway.

This should give us a strong second line of defense against breakages.