Increase zk timeout; clean up some other logic
100ms or 1s is way too low - I just observed a zookeeper fsync for
1.4s. That may be another issue, but it looks like storm sets 20s
by default, so I think we're in good company with 10s. There's not
really a downside to having a larger timeout, in any case.

This also tweaks some logic; it changes to the reconnect logic such
that it doesn't bother trying to reestablish the session, and it
ensures that we don't try to set watches while we're reconnecting.