Implement staged proxying
This is an attempt to try and minimize variance in latency without
increasing load, by introducing a technique in proxying that I'm sure
has a real name, but which I'm calling staged proxying.

The basic idea is this. When we need to proxy a request, and have N
peers which could service that request, we can try concurrent
requests if the peers we try first are being particularly slow. We
use the following algorithm to dispatch requests:

 - Each interval of 'proxy_stage_timeout', starting immediately, a
   request is kicked off to one random not-yet-tried peer. All
   requests after the first run concurrently.
 - If a request finishes successfully to any peer, the result is
   returned immediately and the others are canceled, as long as the
   result was not an error.
 - If a request finishes, but resulted in an error, another is kicked
   off immediately to another random not-yet-tried peer.
 - This process continues until either all peers are being tried, in
   which case the code just waits for one to finish. If the total
   'proxy_timeout' is hit at any point, the method returns
   immediately with an error and cancels any running requests.

There are also some config changes to support this. Since
'proxy_timeout' now has a semantically different meaning, and I moved
a few config properties around, this is a backwards incompatible
change.