Fix a race when removing watches (for real this time)
https://github.com/stripe/sequins/pull/44 was an attempt to fix a
race condition when removing a watch, but the fix was wrong, as this
example demonstrates: https://play.golang.org/p/doj48elk2B

This is a real fix, whereby we let the hookWatchChildren loop close
the client channels itself, but only if it's exiting for the last
time.

This time around, I added a test for the behavior, too.