Stream values to clients
Before this change, we would buffer values into memory and then dump
them out to the client. This is obviously not great for performance.

This fixes that, so we stream values either off disk or from peers.
In the process, it cleans up some lurking issues with proxying:

 - Headers like If-Not-Modified weren't being passed back to peers
 - The X-Sequins-Proxied-To header wasn't always set correctly
 - The X-Sequins-Version header should be set to the peer's version,
   even if it's not the one we asked for
 - Proxied URLs were minorly vulnerable to injection

Unfortunately, this means we can't easily do range queries. This
removes that functionality rather than hack around it, as well as
removing functionality for If-Not-Modified (which we could add back
later with some more comprehensive thought towards caching and Etags).

We also need to enable parallelism for disk reads, or a slow client
could lock a block for a long time reading bytes off lazily. To fix
this, rather than locking the block for every request, we take
advantage of the fact that sparkey files are safe for concurrent
reads. We use a sync.Pool to amortize the memory overhead of these
readers, since each one has a snappy buffer, and that can add up real
fast.