Revert to using timer-based cancellation for proxied requests
In c6eb0b, we switched to using context to manage proxying requests to
peers and the timeouts therein.

Using context still makes this flow easier, but using a context with a
deadline had the unintended side effect that it acts as a deadline for
the entire request, not just to connect. Since we return the still-
open http.Response from the proxy method, we want to leave it open
indefinitely from that point on.

In this way, it was also inadvertently working as a server timeout; we
probably want one of those, but that should be addressed separately
(and not be exclusive to proxied requests).