Completely decouple the building flow from the upgrade flow
By divorcing the process of building new versions from upgrading to
them, we can vastly simplify much of the logic. Versions simply build
asynchronously as soon as they can; serialization is maintained by
having them grab a lock on the DB. The db object, on the other hand,
doesn't keep track of what is building; it simply switches to versions
when they have a complete partition set locally and on peers.

Besides simplifying much of the code, this handles some weird edge-
cases we weren't before:

 - A version can still be switched to even if it's not built locally
   yet (if the peers have a complete partition set). We had some code
   special casing that at startup, but now it works during normal
   upgrade flows as well.

 - During startup, builds weren't respecting max_parallel_loads, which
   meant we could thrash the disk if a node joined a full cluster with
   nothing cached locally. Now that there's only one path for build
   code, they do.