Fix a bug where 404s are turned into 500s
In 38274f, we made the BlockStore object a lot more dumb; in
particular, we removed any knowledge it has about which partitions are
local (that state is instead stored on the partitions object now).

That means that if we don't have the partition in the block store,
it's probably because we never got any keys for that partition, not
anything broken. This might happen if, for instance, there was a file
with no keys in the input data.