Close the body of any outstanding requests.
Even if you cancel a request, it's still the callers responsibility to
close the response Body if `http.Client.Do` returns a non-nil error.
Cancelling a request while it is in the process of streaming a response
merely prevents new data from being added to the response Body, but the
Body itself remains open.